refactor: split search handlers + README refactor + query sparse dispatch
T1-2: Extract batch_search into search/batch.rs and multi_query_search
into search/multi.rs. search/mod.rs now contains only search,
text_search, hybrid_search, search_ids.

T2-2: Refactor README.md from 1614 to 587 lines (64% reduction).
Extract API Reference to docs/reference/API_REFERENCE.md and Business
Scenarios to docs/guides/BUSINESS_SCENARIOS.md. No content lost.

T3-3: Extract sparse query dispatch from query/mod.rs (520 NLOC) into
query/sparse_dispatch.rs (106 NLOC). mod.rs now 418 NLOC.
