+++ crates/velesdb-server/src/handlers/search/mod.rs
@@ -1,5 +1,7 @@
 //! Search handlers for vector similarity, text, and hybrid search.
 
+pub(crate) mod batch;
+pub(crate) mod multi;
 mod pipeline;
 
 use axum::{
@@ -11,9 +13,8 @@ use axum::{
 use std::sync::Arc;
 
 use crate::types::{
-    BatchSearchRequest, BatchSearchResponse, ErrorResponse, HybridSearchRequest,
-    MultiQuerySearchRequest, SearchIdsResponse, SearchRequest, SearchResponse,
-    SearchResultResponse, TextSearchRequest,
+    ErrorResponse, HybridSearchRequest, SearchIdsResponse, SearchRequest, SearchResponse,
+    TextSearchRequest,
 };
 use crate::AppState;
 
@@ -22,6 +23,13 @@ use pipeline::{
     actionable_search_error, build_search_response, execute_search_request, finish_search,
     finish_search_ids, parse_filter_or_400, validate_query_dimension,
 };
+
+#[allow(unused_imports)]
+pub use batch::__path_batch_search;
+pub use batch::batch_search;
+#[allow(unused_imports)]
+pub use multi::__path_multi_query_search;
+pub use multi::multi_query_search;
 
 /// Search for similar vectors.
 ///
@@ -63,238 +71,6 @@ pub async fn search(
     };
 
     finish_search(&state, &name, start, search_result)
-}
-
-/// Batch search for multiple vectors.
-#[utoipa::path(
-    post,
-    path = "/collections/{name}/search/batch",
-    tag = "search",
-    params(
-        ("name" = String, Path, description = "Collection name")
-    ),
-    request_body = BatchSearchRequest,
-    responses(
-        (status = 200, description = "Batch search results", body = BatchSearchResponse),
-        (status = 404, description = "Collection not found", body = ErrorResponse),
-        (status = 400, description = "Invalid request", body = ErrorResponse)
-    )
-)]
-#[allow(clippy::unused_async)]
-pub async fn batch_search(
-    State(state): State<Arc<AppState>>,
-    Path(name): Path<String>,
-    Json(req): Json<BatchSearchRequest>,
-) -> impl IntoResponse {
-    let start = std::time::Instant::now();
-    state.onboarding_metrics.record_search_request();
-
-    let collection = match get_vector_collection_or_404(&state, &name) {
-        Ok(c) => c,
-        Err(resp) => return resp,
-    };
-
-    if let Err(resp) = validate_batch_dimensions(&state, &name, &collection, &req) {
-        return resp;
-    }
-
-    let filters = match parse_batch_filters(&state, &req) {
-        Ok(f) => f,
-        Err(resp) => return resp,
-    };
-
-    let queries: Vec<&[f32]> = req.searches.iter().map(|s| s.vector.as_slice()).collect();
-    let max_top_k = req.searches.iter().map(|s| s.top_k).max().unwrap_or(10);
-
-    let all_results = match collection.search_batch_with_filters(&queries, max_top_k, &filters) {
-        Ok(batch_results) => build_batch_responses(&state, batch_results, &req),
-        Err(e) => {
-            return (StatusCode::BAD_REQUEST, Json(actionable_search_error(&e))).into_response()
-        }
-    };
-
-    let timing_ms = start.elapsed().as_secs_f64() * 1000.0;
-    let duration_us = start.elapsed().as_micros();
-    #[allow(clippy::cast_possible_truncation)]
-    state
-        .db
-        .notify_query(&name, duration_us.min(u128::from(u64::MAX)) as u64);
-
-    Json(BatchSearchResponse {
-        results: all_results,
-        timing_ms,
-    })
-    .into_response()
-}
-
-/// Validate that every query vector in a batch request matches the collection dimension.
-#[allow(clippy::result_large_err)]
-fn validate_batch_dimensions(
-    state: &AppState,
-    name: &str,
-    collection: &velesdb_core::collection::VectorCollection,
-    req: &BatchSearchRequest,
-) -> Result<(), axum::response::Response> {
-    let expected_dimension = collection.config().dimension;
-    for (idx, search) in req.searches.iter().enumerate() {
-        if let Err(error) =
-            validate_query_dimension(state, name, expected_dimension, &search.vector)
-        {
-            return Err((
-                StatusCode::BAD_REQUEST,
-                Json(ErrorResponse {
-                    error: format!("Invalid query at index {idx}: {}", error.error),
-                }),
-            )
-                .into_response());
-        }
-    }
-    Ok(())
-}
-
-/// Parse filters from each search in a batch request.
-#[allow(clippy::result_large_err)]
-fn parse_batch_filters(
-    state: &AppState,
-    req: &BatchSearchRequest,
-) -> Result<Vec<Option<velesdb_core::Filter>>, axum::response::Response> {
-    let mut filters: Vec<Option<velesdb_core::Filter>> = Vec::with_capacity(req.searches.len());
-    for (idx, search) in req.searches.iter().enumerate() {
-        if let Some(filter_json) = &search.filter {
-            match serde_json::from_value(filter_json.clone()) {
-                Ok(filter) => filters.push(Some(filter)),
-                Err(e) => {
-                    state.onboarding_metrics.record_filter_parse_error();
-                    return Err((
-                        StatusCode::BAD_REQUEST,
-                        Json(ErrorResponse {
-                            error: format!(
-                                "Invalid filter at index {idx}: {e}. Hint: validate filter syntax and start with a broader query before reintroducing strict filters."
-                            ),
-                        }),
-                    )
-                        .into_response());
-                }
-            }
-        } else {
-            filters.push(None);
-        }
-    }
-    Ok(filters)
-}
-
-/// Convert batch search results into response objects, recording metrics for empty results.
-fn build_batch_responses(
-    state: &AppState,
-    batch_results: Vec<Vec<velesdb_core::SearchResult>>,
-    req: &BatchSearchRequest,
-) -> Vec<SearchResponse> {
-    let empty_count = batch_results
-        .iter()
-        .filter(|results| results.is_empty())
-        .count();
-    for _ in 0..empty_count {
-        state.onboarding_metrics.record_empty_search_results();
-    }
-    debug_assert_eq!(
-        batch_results.len(),
-        req.searches.len(),
-        "search_batch_with_filters must return one result-vec per query"
-    );
-    batch_results
-        .into_iter()
-        .zip(req.searches.iter())
-        .map(|(results, search)| SearchResponse {
-            results: results
-                .into_iter()
-                .take(search.top_k)
-                .map(|r| SearchResultResponse {
-                    id: r.point.id,
-                    score: r.score,
-                    payload: r.point.payload,
-                })
-                .collect(),
-        })
-        .collect()
-}
-
-/// Multi-query search with fusion strategies.
-#[utoipa::path(
-    post,
-    path = "/collections/{name}/search/multi",
-    tag = "search",
-    params(("name" = String, Path, description = "Collection name")),
-    request_body = MultiQuerySearchRequest,
-    responses(
-        (status = 200, description = "Multi-query search results", body = SearchResponse),
-        (status = 404, description = "Collection not found", body = ErrorResponse),
-        (status = 500, description = "Internal server error", body = ErrorResponse)
-    )
-)]
-#[allow(clippy::unused_async)]
-pub async fn multi_query_search(
-    State(state): State<Arc<AppState>>,
-    Path(name): Path<String>,
-    Json(req): Json<MultiQuerySearchRequest>,
-) -> impl IntoResponse {
-    use velesdb_core::FusionStrategy;
-    state.onboarding_metrics.record_search_request();
-
-    let collection = match get_vector_collection_or_404(&state, &name) {
-        Ok(c) => c,
-        Err(resp) => return resp,
-    };
-
-    let strategy = match req.strategy.to_lowercase().as_str() {
-        "average" | "avg" => FusionStrategy::Average,
-        "maximum" | "max" => FusionStrategy::Maximum,
-        "rrf" => FusionStrategy::RRF { k: req.rrf_k },
-        "weighted" => FusionStrategy::Weighted {
-            avg_weight: req.avg_weight,
-            max_weight: req.max_weight,
-            hit_weight: req.hit_weight,
-        },
-        _ => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(ErrorResponse {
-                    error: format!(
-                        "Invalid strategy: {}. Valid: average, maximum, rrf, weighted",
-                        req.strategy
-                    ),
-                }),
-            )
-                .into_response()
-        }
-    };
-
-    let expected_dimension = collection.config().dimension;
-    for (idx, vector) in req.vectors.iter().enumerate() {
-        if let Err(error) = validate_query_dimension(&state, &name, expected_dimension, vector) {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(ErrorResponse {
-                    error: format!("Invalid query vector at index {idx}: {}", error.error),
-                }),
-            )
-                .into_response();
-        }
-    }
-
-    let query_refs: Vec<&[f32]> = req.vectors.iter().map(Vec::as_slice).collect();
-
-    let results = match collection.multi_query_search(&query_refs, req.top_k, strategy, None) {
-        Ok(r) => r,
-        Err(e) => {
-            return (StatusCode::BAD_REQUEST, Json(actionable_search_error(&e))).into_response()
-        }
-    };
-
-    if results.is_empty() {
-        state.onboarding_metrics.record_empty_search_results();
-    }
-
-    Json(build_search_response(results)).into_response()
 }
 
 /// Search using BM25 full-text search.
